requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
pymysql==1.0.2

drf-yasg==1.20.0
djangorestframework==3.12.4